LayoutTests/webanimations/no-scheduling-while-filling-non-accelerated.html

<!DOCTYPE html>
<html>
<head>
<style>

#target {
    width: 100px;
    height: 100px;
    background-color: black;
}

</style>
</head>
<body>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/rendering-frames.js"></script>
<div id="target"></div>
<script>

promise_test(async () => {
    const keyframes = { marginLeft: ["50px", "100px"] };
    const timing = { duration: 50, fill: "forwards" };

    await document.getElementById("target").animate(keyframes, timing).finished;

    const numberOfTimelineInvalidationsWhenFinished = internals.numberOfAnimationTimelineInvalidations();
    await renderingFrames(2);
    assert_equals(internals.numberOfAnimationTimelineInvalidations(), numberOfTimelineInvalidationsWhenFinished);
}, "There should not be any updates made to the timeline after a forward-filling animation completes.");

</script>
</body>
</html>
